requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
twilio
